requirements.txt
firecrawl-py
python-dotenv
argparse